pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spektci"
version = "0.1.0"
description = "Multi-platform CI/CD compliance scanner for GitHub Actions, Jenkins, Bitbucket, Azure DevOps & CircleCI"
readme = "README.md"
license = { text = "MPL-2.0" }
requires-python = ">=3.11"
authors = [{ name = "spektci contributors" }]
keywords = ["ci", "cd", "compliance", "security", "scanner", "github-actions", "jenkins", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1,<9",
    "rich>=13.0,<14",
    "pydantic>=2.0,<3",
    "pyyaml>=6.0,<7",
    "ruamel.yaml>=0.18,<1",
    "httpx>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "mypy>=1.10,<2",
    "ruff>=0.4,<1",
    "types-PyYAML>=6.0",
    "respx>=0.21,<1",
]

[project.scripts]
spektci = "spektci.cli.main:cli"

[project.urls]
Homepage = "https://github.com/spektci/spektci"
Repository = "https://github.com/spektci/spektci"
Issues = "https://github.com/spektci/spektci/issues"
Documentation = "https://github.com/spektci/spektci/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/spektci"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["ruamel.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["spektci"]

[tool.coverage.run]
source = ["spektci"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
